more refactored
